Ugly but memoized - much faster
We did some quick-and-dirty testing: each time anagrams_for was called,
we put the phrase it was called with to STDOUT, captured those to a
file, sort | uniq -c | sort -nr to see how many times it was called with
each phrase. It was called a LOT with the shortest phrases, and nearly
all phrases were passed in more than once. So memoizing seemed like a
big possible performance improvement, even given that our algo was
correct (like with fibonacci generation).

Yep! It helped a lot. Can now work with much longer phrases than before.
(It also helps tremendously to use a dictionary that doesn't contain
every letter of the alphabet as a "word" - "a" is a word but "r" isn't.)