Simpler and faster - do not produce duplicate anagrams
Previously, we produced a lot of duplicate anagrams, but compensated in
a couple of ways. First, we removed them by using mapsets, so that they
didn't show up in the final results. Second, we used an agent so that if
we were trying to produce anagrams for a phrase we'd already done, we
didn't have to do all the work again.

However, Jared pointed out the first issue, and it made the second clear
as another compromise. The real problem was that we were producing the
duplicates at all.

This happened as follows: say we had the phrase "abc" and the dictionary
["a", "b", "c"].  First, we'd take "a" and find all the anagrams of
"abc" we could build with it and other dictionary entries. That would
include the phrase "a b c". At this point, we'd have every possible answer
that used "a". Having finished a round of iteration with "a", we'd
proceed to "b". What can we build with "b" and the dictionary ["a",
"b", "c"]? Since we've left "a" in the dictionary, we'll build "a b c"
again. But here's the key: *we don't need "a" in the dictionary anymore*.
We've already built every possible anagram that includes "a".

Instead, we throw it out. We switch from a list comprehension to an
Enum.reduce, and part of the state we carry along to the next iteration
is what our current dictionary is. When we start looking for anagrams
that include "b", we use a dictionary that doesn't have "a" in it. Each
new iteration has a smaller dictionary. This avoids building all the
duplicates in the first place, so MapSets aren't needed anymore.

The Agent also isn't needed anymore for caching, since we never try to
find anagrams for the same phrase twice.

Big speedup!

...Ok, "how big is the speedup, Nathan?" is what you're wondering. The
answer is that I was too lazy to properly measure. I got ~6s on the
"slow" test case, vs ~15s on master. If I change that test case to have
a longer phrase, both versions are still too slow to finish before
ExUnit kills them, and ain't nobody got time to wait upwards of 2
minutes to compare.